Bots_Server/requirements.txt
h2>=4.1.0
hyper>=0.7.0
httpx[http2]>=0.24.0
aiohttp>=3.8.0
asyncio
uvloop
flask>=2.3.0
hypercorn>=0.14.0
pyyaml>=6.0
numpy>=1.21.0
pandas>=1.5.0
psutil>=5.9.0
colorlog>=6.7.0
